modify all the interfaces

+++ src/main/java/com/thoughtworks/rslist/api/UserController.java
@@ -3,6 +3,7 @@ package com.thoughtworks.rslist.api;
 import com.thoughtworks.rslist.domain.User;
 import com.thoughtworks.rslist.exception.Error;
 import com.thoughtworks.rslist.exception.RsEventNotValidException;
+import com.thoughtworks.rslist.po.RsEventPO;
 import com.thoughtworks.rslist.po.UserPO;
 import com.thoughtworks.rslist.repository.RsEventRepository;
 import com.thoughtworks.rslist.repository.UserRepository;
@@ -18,7 +19,7 @@ import java.util.Optional;
 
 @RestController
 public class UserController{
-    List<User> userList = new ArrayList<>();
+    List<UserPO> userList = new ArrayList<>();
     @Autowired
     UserRepository userRepository;
     @Autowired
@@ -37,21 +38,29 @@ public class UserController{
     }
 
     @GetMapping("/user/{index}")
-    public ResponseEntity getOneIndexUser(@RequestBody @Valid int index){
-        userRepository.deleteById(index);
-        return ResponseEntity.ok().build();
+    public ResponseEntity getOneIndexUserInfor(@RequestBody @Valid int index){
+        Optional<UserPO> user = userRepository.findById(index);
+        if(user.get() != null){
+            return ResponseEntity.ok(user.get());
+        }
+        return ResponseEntity.badRequest().build();
+
     }
 
-    @GetMapping("/user")
-    public List<User> getUserList(){
-        return userList;
+    @GetMapping("/user/list")
+    public ResponseEntity getUserList(){
+        userList = userRepository.findAll();
+        return ResponseEntity.ok(userList);
     }
 
     @DeleteMapping("/user/{index}")
     public ResponseEntity deleteUser(@RequestBody @Valid int index){
-        Optional <UserPO>  user = userRepository.findById(index);
-        userRepository.deleteById(index);
-        return ResponseEntity.ok().build();
+        if(userRepository.findById(index) != null){
+            userRepository.deleteById(index);
+            return ResponseEntity.ok().build();
+        }else{
+            return ResponseEntity.badRequest().build();
+        }
     }
 
     @ExceptionHandler({RsEventNotValidException.class, MethodArgumentNotValidException.class})
